Better organize tests
I've moved integration tests into a `tests` folder. I've also ignored
the builder integration test since it's still incredibly slow.
