+++ src/builder.rs
@@ -8,15 +8,3 @@ pub fn build_complete_puzzle() -> Vec<Vec<isize>> {
 
     solver.solve_puzzle()
 }
-
-#[cfg(test)]
-mod tests {
-    use validator::valid;
-    use builder::build_complete_puzzle;
-
-    #[test]
-    fn it_builds_a_valid_puzzle() {
-        let puzzle = build_complete_puzzle();
-        assert!(valid(&puzzle));
-    }
-}
